SchoolProj/requirements.txt
pygame
loguru
colorama
mysql-connector
